fix: async OAuth processing and session validation logic
- Fix critical session validation logic error (was calling async on success instead of when needed)
- Add timeout controls (30s) to async operations to prevent goroutine leaks
- Improve async flow control with proper default case handling
- Extract API key handlers to separate file for better organization
- Move sanitizePathForLogging to logger.go where it belongs
- Add IsNeedValidateSession optimization to avoid unnecessary validations

These changes prevent Envoy worker thread deadlocks by processing OAuth operations asynchronously.
